Moves individual decorators into subpackages. We have:
1. configuration -- if/else equivalents
2. dependencies -- utilities for specifying dependencies, generally used
   for parameterized, but this will likely be expanded. Note decorators.
3. expanders -- decorators that turn one node into many
4. function_modifiers_base -- the base classes/common utilities
5. macros -- decorators that replace a function's implementation with
   something else (think @does/@dynamic_transform)
6. metadata -- decorators that add metadata to a node (tag right now)
7. validation -- decorators that do some validation on a node

Note that this was all done to preserve pure backwards compatibility.
Care was taken that every public component of the API was placed within
the __init__.py of function_modifiers.
