+++ hamilton/function_modifiers/configuration.py
@@ -1,0 +1,103 @@
+from typing import Any, Callable, Collection, Dict
+
+from . import function_modifiers_base
+
+"""Decorators that handle the configuration of a function. These can be viewed as
+replacing if/else/switch statements in standard dataflow definition libraries"""
+
+
+class config(function_modifiers_base.NodeResolver):
+    """Decorator class that resolves a node's function based on  some configuration variable
+    Currently, functions that exist in all configurations have to be disjoint.
+    E.G. for every config.when(), you can have a config.when_not() that filters the opposite.
+    That said, you can have functions that *only* exist in certain configurations without worrying about it.
+    """
+
+    def __init__(self, resolves: Callable[[Dict[str, Any]], bool], target_name: str = None):
+        self.does_resolve = resolves
+        self.target_name = target_name
+
+    def _get_function_name(self, fn: Callable) -> str:
+        if self.target_name is not None:
+            return self.target_name
+        return function_modifiers_base.sanitize_function_name(fn.__name__)
+
+    def resolve(self, fn, configuration: Dict[str, Any]) -> Callable:
+        if not self.does_resolve(configuration):
+            return None
+        fn.__name__ = self._get_function_name(fn)  # TODO -- copy function to not mutate it
+        return fn
+
+    def validate(self, fn):
+        if fn.__name__.endswith("__"):
+            raise function_modifiers_base.InvalidDecoratorException(
+                "Config will always use the portion of the function name before the last __. For example, signups__v2 will map to signups, whereas"
+            )
+
+    @staticmethod
+    def when(name=None, **key_value_pairs) -> "config":
+        """Yields a decorator that resolves the function if all keys in the config are equal to the corresponding value
+
+        :param key_value_pairs: Keys and corresponding values to look up in the config
+        :return: a configuration decorator
+        """
+
+        def resolves(configuration: Dict[str, Any]) -> bool:
+            return all(value == configuration.get(key) for key, value in key_value_pairs.items())
+
+        return config(resolves, target_name=name)
+
+    @staticmethod
+    def when_not(name=None, **key_value_pairs: Any) -> "config":
+        """Yields a decorator that resolves the function if none keys in the config are equal to the corresponding value
+
+        :param key_value_pairs: Keys and corresponding values to look up in the config
+        :return: a configuration decorator
+        """
+
+        def resolves(configuration: Dict[str, Any]) -> bool:
+            return all(value != configuration.get(key) for key, value in key_value_pairs.items())
+
+        return config(resolves, target_name=name)
+
+    @staticmethod
+    def when_in(name=None, **key_value_group_pairs: Collection[Any]) -> "config":
+        """Yields a decorator that resolves the function if all of the keys are equal to one of items in the list of values.
+
+        :param key_value_group_pairs: pairs of key-value mappings where the value is a list of possible values
+        :return: a configuration decorator
+        """
+
+        def resolves(configuration: Dict[str, Any]) -> bool:
+            return all(
+                configuration.get(key) in value for key, value in key_value_group_pairs.items()
+            )
+
+        return config(resolves, target_name=name)
+
+    @staticmethod
+    def when_not_in(**key_value_group_pairs: Collection[Any]) -> "config":
+        """Yields a decorator that resolves the function only if none of the keys are in the list of values.
+
+        :param key_value_group_pairs: pairs of key-value mappings where the value is a list of possible values
+        :return: a configuration decorator
+
+        :Example:
+
+        @config.when_not_in(business_line=["mens","kids"], region=["uk"])
+        def LEAD_LOG_BASS_MODEL_TIMES_TREND(TREND_BSTS_WOMENS_ACQUISITIONS: pd.Series,
+                                    LEAD_LOG_BASS_MODEL_SIGNUPS_NON_REFERRAL: pd.Series) -> pd.Series:
+
+        above will resolve for config has {"business_line": "womens", "region": "us"},
+        but not for configs that have {"business_line": "mens", "region": "us"}, {"business_line": "kids", "region": "us"},
+        or {"region": "uk"}
+
+        .. seealso:: when_not
+        """
+
+        def resolves(configuration: Dict[str, Any]) -> bool:
+            return all(
+                configuration.get(key) not in value for key, value in key_value_group_pairs.items()
+            )
+
+        return config(resolves)
